Fixes and Features
I fixed the bugs and missing features with the game over menu. Player Movement is now frozen when the game ends and everything stops spawning. I also added a pause menu which freezes everything in the game and allows users to resume, return to the main menu, or quit. I also added a main menu button to the game over menu.